requirements.txt
idx2numpy==1.2.3
matplotlib==3.4.3
numpy==1.22.4
opencv_python==4.6.0.66
Pillow==9.4.0
scikit_learn==1.2.2
tensorflow==2.10.0
